Set up descriptor layout / pool / sets to use the standardised set indices
This is a first, very rough, draft of defining a group of descriptor sets that can be standardised across the shaders. With this method we will load certain sets to certain indicies, such as the view data, which won't change per primitive. Later when we have multiple pipelines we can easily use the data across them while knowing we can bind whatever we want into set 0 for example