Removed EntryPoint Application is now created through friend functions
This is to hopefully allow more control over how the application is created an initialised. The user currently has a choice between a windowed or windowless application.

The current trade off with this approach is now the main loop is client side and we don't have control over it any more. But this could be a good thing