Manually record each metal render command into a function pointer
This primarily acts as a work around for the fact that all command buffers need to recorded and committed to a queue in metal before presenting the drawable. This method also comes with the added side effect of recording a buffer once and then reusing it later multiple times. Which is impossible in metal as once commit is called on a buffer it can't be reused