Removed support for arrays of textures in descriptor sets
This is to be more compatible with how Metal handles arrays of textures. Each element in the array is unrolled and has a unique ID causing the IDs in the descriptor sets to be invalid. Texture arrays do not work like this in Metal, keeping the binding IDs consistent across all APIs