Refactored ListAllocator to stop storing headers in the backing memory
This will allow the full size of the backing memory to be used in allocations which is the expected usage of this allocator. It also makes the allocator a bit more robust in general as there no longer needs to be a lot of pointer arithmetic to compensate for the header offset.