Removed the `RgImageView` object.
Handling all the views and how they overlap was too complicated. This was mainly done to as a way to provide manual layout transitions for any image for undefined -> used. However with the fact that image layouts can now be explicitly specified when creating them this shouldn't be required anymore. Images can be created in their fall back layout and in most cases barriers shouldn't need to be inserted